massive re-organization of the code, preparation for actual testing
