+++ .gitignore
@@ -1,5 +1,5 @@
 __pycache__
 *.pyc
-/env/
+/*env*/
 .*
 *.egg-info
