+++ better_test/harness/crash.py
@@ -1,0 +1,7 @@
+from django.utils import unittest
+
+
+class Tests(unittest.TestCase):
+    def test_segfault(self):
+        from segfault import segfault
+        segfault()
