+++ better_test/tests/test_crash.py
@@ -1,0 +1,6 @@
+from django.utils import unittest
+
+
+class Tests(unittest.TestCase):
+    def test_segfault(self):
+        pass
