+++ test_requirements.py
@@ -1,0 +1,1 @@
+https://github.com/ojii/segfault/archive/master.tar.gz#egg=segfault
